[iOS][core] Migrate some unit tests to Swift Testing (#42662)
